used socked.io for recieving webhook info for real time updates.

+++ api/src/utils/imageProcessor.js
@@ -2,7 +2,7 @@ import sharp from 'sharp'; // For image processing
 import axios from 'axios'; // To download images
 import fs from 'fs';
 import path from 'path';
-// import RequestModel from '../models/requestModel.js'; // Import the Request model for updates
+import Logger from '../utils/logger.js';
 
 export const processImages = async (inputImageUrls) => {
     const outputImageUrls = [];
@@ -10,8 +10,18 @@ export const processImages = async (inputImageUrls) => {
     try {
         // Process each image URL by downloading, compressing, and saving the output
         for (const url of inputImageUrls) {
-            const imagePath = path.join('uploads', path.basename(url)); // Path to save the downloaded image
+            const imageName = path.basename(url);
+            const imagePath = path.join('uploads', imageName); // Path to save the downloaded image
             
+            // Check if the image already exists in the uploads folder
+            if (fs.existsSync(imagePath)) {
+                Logger.info(`Image already exists: ${imagePath}. Skipping processing.`);
+                // Optionally, you can return the existing URL
+                const existingImageUrl = `http://localhost:3000/uploads/${imageName}`;
+                outputImageUrls.push(existingImageUrl);
+                continue; // Skip to the next image URL
+            }
+
             // Download the image
             const response = await axios.get(url, { responseType: 'arraybuffer' });
             fs.writeFileSync(imagePath, response.data); // Save the image locally
@@ -22,7 +32,7 @@ export const processImages = async (inputImageUrls) => {
                 .jpeg({ quality: 50 }) // Compress image
                 .toBuffer();
 
-            const outputFileName = `compressed-${path.basename(url)}`; // Use the basename of the input URL
+            const outputFileName = `compressed-${imageName}`; // Use the basename of the input URL
             const outputPath = path.join('uploads', outputFileName);
             await sharp(outputBuffer).toFile(outputPath);
 
@@ -31,11 +41,11 @@ export const processImages = async (inputImageUrls) => {
             outputImageUrls.push(outputImageUrl); // Store the public URL of the compressed image
         }
 
-        console.log("Output image urls: " + JSON.stringify(outputImageUrls));
+        Logger.info("Output image urls: " + JSON.stringify(outputImageUrls));
         return outputImageUrls; // Return the processed output image URLs
 
     } catch (error) {
-        console.error('Error processing images:', error);
+        Logger.error('Error processing images:', error);
         throw error; // Rethrow error for handling in the calling function
     }
 };
